Add the main file and the PyCharm configuration.
I use the PyCharm editor for this project. I added a project
configuration to be able to start the simulator and the training.
Another possibility is to run this command in a terminal from the root
folder of the project:
python -m main --parameters <path to root folder of pypownet>/pypownet/parameters/default14 --niter 1000 --verbose --render